moved type definitions
