+++ client/src/types/volunteer.ts
@@ -1,0 +1,11 @@
+export interface Volunteer {
+  id: number;
+  firstName: string;
+  lastName: string;
+  email: string;
+  phoneNumber?: string;
+  role?: string;
+  specializations?: string[];
+  languages?: string[];
+  experienceLevel?: string;
+}
